:green_heart: Extract tokenizer_test_base.go
